Fixing an issue where portfolio on mobile
The portfolio page is not rendering properly on mobile when directly accessing the page via url or refreshing the page. This is a github page specific issue

I think currently it only contains portfolio._index. And portfolio.tsx is missing.
The logic checks whether a logo exists. If not, it forcibly includes the portfolioLayoutWidget in the page.